requirements/local_main.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
defusedxml==0.7.1
Django==4.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
httpie==3.1.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
